Move the default logoutFromAuthServer to the utils module
