+++ packages/gatekeeper/src/components/logout.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 import { LOGIN_URL } from '../helpers/constants';
+import { logoutFromAuthServer } from '../helpers/utils';
 
 /** interface to describe props for Logout component
  * @member {typeof logOutUser}logoutActionCreator action creator that logs out user.
@@ -24,22 +25,6 @@ export const defaultLogoutProps: LogoutProps = {
   logoutURL: null,
   redirectPath: LOGIN_URL
 };
-
-/**
- * Open another window and navigate to the logout URL.
- * @param {string} logoutURL URL string representing the auth server logout URL endpoint.
- * This function takes the approach of opening a new window and navigating to the logout
- * url of the authentication server in order to go around the browser's CORS policy.
- */
-export function logoutFromAuthServer(logoutURL: string) {
-  const logoutWindow: Window | null = window.open(logoutURL);
-  const timer: NodeJS.Timeout = setInterval(() => {
-    if (logoutWindow) {
-      logoutWindow.close();
-    }
-    clearInterval(timer);
-  }, 20);
-}
 
 /** Logout component */
 const Logout = (props: LogoutProps) => {
